GameActivity.kt size gameElements calculated from screen size

+++ app/src/main/java/com/chico/sapper/Metrics.kt
@@ -4,6 +4,6 @@ class Metrics (){
     var sizeDisplayX:Int = 0
     var sizeDisplayY:Int = 0
 
-    var gameCellSize:Int = 0
+    var gameCellSize:Double = 0.0
 
 }
